+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/HashAlgorithm/HashAlgorithmIntegrationCommon.cs
@@ -4,11 +4,9 @@
 // </copyright>
 
 using System;
-using System.Linq;
 using Datadog.Trace.Configuration;
 using Datadog.Trace.IAST;
 using Datadog.Trace.Logging;
-using Datadog.Trace.Tagging;
 
 namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.HashAlgorithm
 {
